+++ web-interface/control/www/service/ipbx/asterisk/pbx_settings/users/deletes.php
@@ -32,7 +32,10 @@ for($i = 0;$i < $arr['cnt'];$i++)
 	{
 		if(($info['ufeatures'] = $ufeatures->get_by_protocol($v[$j],$k)) === false
 		|| ($info['protocol'] = $protocol->get($info['ufeatures']['protocolid'])) === false
-		|| ($interface = $ipbx->mk_interface($info['ufeatures']['protocol'],$info['protocol']['name'])) === false)
+		|| ($interface = $ipbx->mk_interface($info['protocol']['name'],
+						     $info['ufeatures']['protocol'],
+			   		  	     $info['ufeatures']['number'],
+			   		  	     $info['protocol']['context'])) === false)
 			continue;
 
 		if($protocol->delete($info['protocol']['id']) === false)
@@ -59,7 +62,8 @@ for($i = 0;$i < $arr['cnt'];$i++)
 			continue;
 		}
 
-		if($qmember->delete_by_interface($interface) === false)
+		if($qmember->get_list_by_interface($interface) !== false
+		&& $qmember->delete_by_interface($interface) === false)
 		{
 			$protocol->add_origin();
 			$ufeatures->add_origin();
@@ -71,9 +75,9 @@ for($i = 0;$i < $arr['cnt'];$i++)
 		}
 
 		$hints_where['exten'] = $info['ufeatures']['number'];
-		$hints_where['app'] = $interface;
 
-		if(($info['extensions'] = $extensions->get_where($hints_where)) !== false)
+		if(($hints_where['app'] = $ipbx->mk_interface($info['protocol']['name'],$info['ufeatures']['protocol'])) !== false
+		&& ($info['extensions'] = $extensions->get_where($hints_where)) !== false)
 			$extensions->delete($info['extensions']['id']);
 
 		if(($info['usergroup'] = $ugroup->get_by_user($info['ufeatures']['id'])) !== false)
